Switch to 1.2.0 clone for bids-examples and fix test hang
This updates the submodule to use a logical name and point at the tag
for 1.2.0. I also updated documentation to default to a shallow clone
which is a bit faster than a full clone of bids-examples.

The hang was caused by .git directories being recursively followed now
that it's a submodule. Filtering .git from collected directories for
tests solves this.

This also simplifies a lot of logic around bids-examples versions now
that it is tracked as a submodule. It makes more sense to give the
directory a logical name without a version and refer to git metadata
for the precise version.

I went ahead and removed the test environments now that we don't need
to inject the version into dynamically generated paths as well. This
eliminates the need to use jest-environment-jsdom-global at all, so
it's been removed.